feat(projects): add CodeOps Studio
Included [CodeOps Studio](https://github.com/euptron/CodeOps-Studio) in the list of featured projects on **madeinnigeria.dev.**

CodeOps Studio is a mobile-based integrated development environment (IDE) and code editor for Android devices. Designed to support developers working on the go, it stands out as a notable open-source project created by a Nigerian developer.

The project was added to Group C and positioned appropriately in alphabetical order.